restructure construct modules
